requirements.txt
appdirs==1.4.4
asttokens==2.2.1
audioread==3.0.0
backcall==0.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.40.0
matplotlib-inline==0.1.6
msgpack==1.0.5
nest-asyncio==1.5.6
numba==0.57.0
numpy==1.24.3
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pooch==1.6.0
praat-parselmouth==0.4.3
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
zipp==3.15.0
